[AOTI][dashboard] Update how peak memory is measured
Summary: In the dashboard measurement script, AOTI needs to run Eager first to register the output pytree, so the peak memory compression ratio on the dashboard is always close to 1. Update AOTI run to use an extra warmup run, so the peak memory compression ratio measures the result at the run time instead of the compile time.

ghstack-source-id: c75e413fef5b8a67e542fab96110650341b98d5d
Pull Request resolved: https://github.com/pytorch/pytorch/pull/150534