Implement needs_exact_strides for mutable custom operators
Mutable custom operators get wrapped into an auto_functionalized HOP, so
we need to store the arg_kwarg_vals on the auto_functionalized HOP
itself.

When Inductor does the re-inplacing, it'll use the pattern matcher to
decompose the auto_functionalized HOP back into the original op (and
0+ other view or clone operations). The pattern matcher uses the
arg_kwarg_vals to trace the subgraph to do the decomposition, so it
ultimately sets arg_kwarg_vals on the original op's node correctly.

Test Plan:
- new test

ghstack-source-id: 0f6fa0310557b041469b3ff3b41cd888134acf26
Pull Request resolved: https://github.com/pytorch/pytorch/pull/148091