Add config option to force disable CompiledTritonKernel cache
Pull Request resolved: https://github.com/pytorch/pytorch/pull/150778

We're unfortunately still seeing some flakiness internally in specific internal models: adding a disable CompiledTritonKernels cache feature to help mitigate.

The issue seems to be sucluded to this specific model: StaticCudaLauncher could also help alleviate it, though I haven't had the permissions to be able to test yet. It's unclear to me if this will definitively fix the issue for the job, but we can test and if it doesn't, we'll have removed another possible cause.

Differential Revision: [D72584099](https://our.internmc.facebook.com/intern/diff/D72584099/)
ghstack-source-id: 276613091