[standalone_compile] support multiple returns
We were only returning the first one. There's an edge case on what to do
if the original function returns a single Tensor. capture(f) returns a
function that returns a tuple of one Tensor in this case and we were
originally converting this back to one single Tensor. I think it's fine
to return a tuple of one Tensor (that is what the graph passed to
standalone_compile asked for!) but we can revisit.
fine

Test Plan:
- modified one test to used multiple outputs

ghstack-source-id: fb33373404653934335510bc7abe52779daf8461
Pull Request resolved: https://github.com/pytorch/pytorch/pull/151551