[standalone_compile] Some misc fixes
This PR fixes two things.

The first problem is that in the vLLM style standalone_compile is
called from within a custom torch.compile backend. If there already is a
FakeTensorMode (which there is), we shouldn't create a new
FakeTensorMode with the same shape_env, instead we should just reuse the
same FakeTensorMode.

The second thing is that compile_fx can mutate the passed in gm, so we
deepcopy (since standalone_compile should be standalone)

Test Plan:
- new test
- updated old tests

ghstack-source-id: c670cbad1b4b2aff5f3a60e476abe844b9b1dc04
Pull Request resolved: https://github.com/pytorch/pytorch/pull/151502