[logging] Put "everything" WaitCounters in dynamo_timed
Summary: We have a couple WaitCounters that we want to capture all runtime and compile overheads. Put them in dynamo_timed so we don't have to make sure we have them installed at all the outer compile and runtime functions. Also, dynamo_timed already has to figure out if we're timing a runtime vs. compile-time event, so we can reuse some of that logic.

Test Plan:

ghstack-source-id: 803c282f1d6438311d5b09416d8fa5d0c46dad23
Pull Request resolved: https://github.com/pytorch/pytorch/pull/151757