Extend compute_global_tensor_shape to multi dimension sharding
`compute_global_tensor_shape` until all gathers shape of the local
tensors from all the ranks and then computes the shape of the global
DTensor based on the device mesh and the placements. Earlier this util
supported only 1D device mesh, extending the util to now support multi
dimension sharding.

Here we take a recursive approach to calculate the global shape via
a DFS on the device mesh.

ghstack-source-id: deba3456469ef8f8a63d6053096b3efc0e80527b
Pull Request resolved: https://github.com/pytorch/pytorch/pull/152166